Added some code to set some aliases only when cygwin it's running

+++ .bash_aliases
@@ -56,9 +56,11 @@ alias elearning="bash $work_dir/virtual_tours/.archives/bin/elearning/elearning.
 # Cygwin
 # ========================================
 
-alias chrome='/c/Program\ Files\ \(x86\)/Google/Chrome/Application/chrome.exe &'
-alias emacs='/c/emacs-24.2/bin/runemacs.exe &'
-alias halt='shutdown -p -f'
+if [ $(uname -o) = "Cygwin" ];then
+    alias chrome='/c/Program\ Files\ \(x86\)/Google/Chrome/Application/chrome.exe &'
+    alias emacs='/c/emacs-24.2/bin/runemacs.exe &'
+    alias halt='shutdown -p -f'
+fi
 
 # ========================================
 # Server
